+++ pyrogram/types/messages_and_media/__init__.py
@@ -36,8 +36,8 @@ from .venue import Venue
 from .video import Video
 from .video_note import VideoNote
 from .voice import Voice
-from .webpage import WebPage
 from .web_app_data import WebAppData
+from .web_page import WebPage
 
 __all__ = [
     "Animation", "Audio", "Contact", "Document", "Game", "Location", "Message", "MessageEntity", "Photo", "Thumbnail",
